+++ private/system/lib-comment.php
@@ -106,6 +106,7 @@ function plugin_subscription_email_format_comment($category,$track_id,$post_id,$
         $T->set_var(array(
             'post_subject'  => $A['title'],
             'post_date'     => $date,
+            'iso8601_date'  => $dt->toISO8601(),
             'post_name'     => $name,
             'post_comment'  => $A['comment'],
             'notify_msg'    => $notifymsg,
@@ -505,6 +506,8 @@ function CMT_getComment( &$comments, $mode, $type, $order, $delete_option = fals
                     $_CONF['site_url'] . '/users.php?mode=profile&amp;uid=' . $A['uid']
                 )
             );
+            $template->set_var( 'author_url', $_CONF['site_url'] . '/users.php?mode=profile&amp;uid=' . $A['uid']);
+
 
         } else {
             $username = $filter->sanitizeUsername($A['name']);
@@ -514,6 +517,7 @@ function CMT_getComment( &$comments, $mode, $type, $order, $delete_option = fals
             $template->set_var( 'author', $username);
             $template->set_var( 'author_fullname', $username);
             $template->set_var( 'author_link', @htmlspecialchars($username,ENT_COMPAT,COM_getEncodingt() ));
+            $template->unset_var( 'author_url');
 
             if( $_CONF['allow_user_photo'] ) {
                 $template->set_var( 'author_photo_raw', $_CONF['default_photo'] );
@@ -545,6 +549,7 @@ function CMT_getComment( &$comments, $mode, $type, $order, $delete_option = fals
         $dtObject = new Date($A['nice_date'],$_USER['tzid']);
 
         $template->set_var( 'date', $dtObject->format($_CONF['date'],true));
+        $template->set_var( 'iso8601_date', $dtObject->toISO8601() );
         $template->set_var( 'sid', $A['sid'] );
         $template->set_var( 'type', $A['type'] );